requirements.txt
python-telegram-bot>=22.5
psutil>=5.9.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
requests>=2.28.0
beautifulsoup4>=4.12.0
playwright>=1.40.0
lxml>=4.9.0
feedparser>=6.0.10